requirements.txt
asgiref==3.8.1
Brotli==1.1.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
dj-database-url==2.2.0
Django==5.1
django-browser-reload==1.15.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
h11==0.14.0
idna==3.10
packaging==24.1
psycopg==3.2.1
psycopg-binary==3.2.1
psycopg2-binary==2.9.9
PyJWT==2.9.0
python-dotenv==1.0.1
requests==2.32.3
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.31.0
whitenoise==6.7.0
